+++ packages/components/scripts/jobs/cem.js
@@ -1,9 +1,13 @@
-import { execPromise, job } from './shared.js';
+import { execPromise, formatFile, getPath, job } from './shared.js';
 
 /**
- * Run esbuild for the components package
+ * Run cem for the components package
  */
-export const runCem = job('Creating component manifest...', async () => await execPromise(
-  'cem analyze --litelement --outdir dist',
-  { stdio: 'inherit' },
-));
+export const runCem = job('Creating component manifest...', async () => {
+  await execPromise(
+    'cem analyze --litelement --outdir dist',
+    { stdio: 'inherit' },
+  );
+
+  await formatFile(getPath('../package.json'), 'json-stringify');
+});