+++ packages/components/scripts/vendorism/custom/index.js
@@ -2,5 +2,7 @@ export * from './button.vendorism.js';
 export * from './customElements.vendorism.js';
 export * from './icon.vendorism.js';
 export * from './input.vendorism.js';
+export * from './icon-button.vendorism.js';
 export * from './textarea.vendorism.js';
 export * from './webTestRunner.vendorism.js';
+export * from './tag.vendorism.js';