+++ packages/components/src/components/tag/tag.custom.styles.ts
@@ -1,0 +1,103 @@
+import { css } from 'lit';
+
+export default css`
+  .tag {
+    background-color: var(--syn-color-neutral-0);
+    border-color: var(--syn-color-neutral-400);
+    color: var(--syn-input-color);
+  }
+
+  .tag .tag__remove {
+    color: var(--syn-color-neutral-500);
+  }
+
+
+  /* Size small */
+  .tag--small {
+    font-size: var(--syn-font-size-x-small);
+    height: var(--syn-font-size-x-large);
+  }
+
+  .tag--small.tag--removable {
+    padding-inline-end: var(--syn-spacing-2x-small);
+  }
+
+
+  /* Size medium */
+  .tag--medium {
+    font-size: var(--syn-font-size-small);
+    height: var(--syn-font-size-2x-large);
+  }
+
+  .tag--medium.tag--removable {
+    padding-inline-end: var(--syn-spacing-x-small);
+  }
+
+  .tag--medium .tag__remove {
+    margin-inline-start: var(--syn-spacing-small);
+  }
+
+
+  /* Size large */
+  .tag--large {
+    font-size: var(--syn-font-size-medium);
+    height: var(--syn-font-size-3x-large);
+  }
+
+  .tag--large .tag__remove {
+    margin-inline-start: var(--syn-spacing-medium);
+  }
+
+  .tag--large.tag--removable {
+    padding-inline-end: var(--syn-spacing-small);
+  }
+
+
+  /* Hover */
+  .tag.tag--removable:hover {
+    background-color: var(--syn-color-neutral-100);
+  }
+  
+  .tag .tag__remove:hover {
+    color: var(--syn-color-neutral-950);
+  }
+
+
+  /* Focus 
+     ! Focus styling is currently not working because shoelace does not support to focus icon-button 
+  */
+  .tag.tag--removable:focus ~ .tag__remove {
+    border: solid 1px var(--syn-color-primary-500);
+  }
+
+
+  /* Slotted icon style */
+  .tag--small ::slotted(syn-icon) {
+    font-size: var(--syn-font-size-small);
+    margin-inline-end: var(--syn-spacing-2x-small);
+  }
+
+  .tag--medium ::slotted(syn-icon) {
+    font-size: var(--syn-font-size-medium);
+    margin-inline-end: var(--syn-spacing-x-small);
+  }
+
+  .tag--large ::slotted(syn-icon) {
+    font-size: var(--syn-font-size-large);
+    margin-inline-end: var(--syn-spacing-x-small);
+  }
+
+
+  /* Remove icon-button style */
+  .tag--small .tag__remove::part(base){
+    font-size: var(--syn-font-size-medium);
+  }
+
+  .tag--medium .tag__remove::part(base){
+    font-size: var(--syn-font-size-x-large);
+  }
+
+  .tag--large .tag__remove::part(base){
+    font-size: var(--syn-font-size-2x-large);
+  }
+`;